Add own listings format
No got a equation numbering on lists we have to use an own command with
minipages. It uses two params.

\eqlst{filename}{labelname}{caption}

After a usage of the command above you could add a reference to point on
the equation number with \ref{labelname}. There will be *no* captions
anymore for the listings inside the document. The caption is only used
for the listoflistings(if used). If need more you have to add another
params to the command.